move OrderingComponent to ResultComponent
