+++ src/app/shell/main/result/organization-cards-list/ordering/ordering.component.scss
@@ -1,12 +1,19 @@
 .ordering-button {
   position: relative;
   width: 120px;
+  height: 2.125rem;
+  background: #FFFFFF;
+  box-shadow: 0px 6px 16px rgba(0, 0, 0, 0.08);
+  border-radius: 20px;
+  border: none;
+  font-weight: bold;
 }
 .options {
   position: absolute;
   background-color: white;
   border: 1px solid rgba(0,0,0,.12);
   width: 120px;
+  z-index: 1;
   .option {
     cursor: pointer;
     &:hover {
