+++ src/app/shell/main/result/organization-cards-list/organization-cards-list.component.ts
@@ -6,10 +6,10 @@ import {orgCard} from '../../../../shared/models/org-card.model';
 import {getCards, SetOrder} from '../../../../shared/store/filter.actions';
 
 
-export interface Option {
-  value: string;
-  viewValue: string;
-}
+// export interface Option {
+//   value: string;
+//   viewValue: string;
+// }
 
 @Component({
   selector: 'app-organization-cards-list',
@@ -19,16 +19,16 @@ export interface Option {
 export class OrganizationCardsListComponent implements OnInit {
 
   @Select(FilterState.orgCards) orgCards$: Observable<orgCard[]>;
-  @Select(state => state.filter.order) order$: Observable<any>;
+  // @Select(state => state.filter.order) order$: Observable<any>;
 
   public cards: orgCard[];
-  options: Option[] = [
-    {value: 'ratingDesc', viewValue: 'Рейтинг'},
-    {value: 'ratingAsc', viewValue: 'Рейтинг'},
-    {value: 'priceDesc', viewValue: 'Ціна'},
-    {value: 'priceAsc', viewValue: 'Ціна'}
-  ];
-  selectedOption: Option;
+  // options: Option[] = [
+  //   {value: 'ratingDesc', viewValue: 'Рейтинг'},
+  //   {value: 'ratingAsc', viewValue: 'Рейтинг'},
+  //   {value: 'priceDesc', viewValue: 'Ціна'},
+  //   {value: 'priceAsc', viewValue: 'Ціна'}
+  // ];
+  // selectedOption: Option;
 
   constructor(private store: Store) {
   }
@@ -36,17 +36,17 @@ export class OrganizationCardsListComponent implements OnInit {
   ngOnInit(): void {
     this.store.dispatch(new getCards())
     this.orgCards$.subscribe((orgCards: orgCard[]) => this.cards = orgCards)
-    this.order$.subscribe(order => {
-      this.selectedOption = this.options.find(option => option.value === order);
-    });
+    // this.order$.subscribe(order => {
+    //   this.selectedOption = this.options.find(option => option.value === order);
+    // });
   }
 
-  setOrder(order: string) {
-    this.store.dispatch(new SetOrder(order));
-  }
+  // setOrder(order: string) {
+  //   this.store.dispatch(new SetOrder(order));
+  // }
 
-  onChangeOrder(order: string) {
-    this.setOrder(order);
-  }
+  // onChangeOrder(order: string) {
+  //   this.setOrder(order);
+  // }
 
 }
